backend/pyproject.toml
[project]
name = "mesh-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.133.1",
    "httpx>=0.28.1",
    "openai==1.63.2",
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "spacy>=3.8.11",
    "uvicorn>=0.41.0",
]
